requirements.txt
google-generativeai>=0.3.0
jpype1==1.5.0
python-dotenv>=1.0.0
requests>=2.28.0
